+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Spell scrolls now generate with custom names based on spell 
 
 API Changes:
+- Add `SpellEvents.HEAL`
 - Add `SpellTagsNumbered` api to assign numbers to spell tags
 - Spell scrolls can now have custom texture based on tag membership of the first contained spell